% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crits.R
\name{KGE}
\alias{KGE}
\title{Kling-Gupta criteria wrapper}
\usage{
KGE(x, y, type = "2011")
}
\arguments{
\item{x}{Observed series}

\item{y}{Simulated series}

\item{type}{Version of computation}
}
\description{
The
}
