% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crits.R
\name{LME}
\alias{LME}
\title{Liu Mean Efficiency}
\usage{
LME(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\description{
Liu Mean Efficiency (LME) R wrapper function. This criteria is defined as 
\deqn{\text{LME} = 1 - \sqrt{(k_1 - 1)^2 + (\beta - 1)^2}}
}
